+++ src/promised-db.ts
@@ -3,25 +3,47 @@
 // (c) 2016-Present by @zenmumbler
 
 declare global {
-	interface PDBDatabaseInfo {
+	interface IDBDatabaseInfo {
 		name: string;
 		version: number;
 	}
 
 	interface IDBFactory {
-		databases?(): Promise<PDBDatabaseInfo[]>;
+		/** New in IndexedDB 3.0 - not widely implemented */
+		databases(): Promise<IDBDatabaseInfo[]>;
+	}
+
+	type IDBTransactionDurability = "default" | "strict" | "relaxed";
+
+	interface IDBTransactionOptions {
+		durability?: IDBTransactionDurability;
+	}
+
+	interface IDBDatabase {
+		transaction(storeNames: string | string[], mode?: IDBTransactionMode, options?: IDBTransactionOptions): IDBTransaction;
+	}
+
+	interface IDBTransaction {
+		/** New in IndexedDB 3.0 - not widely implemented */
+		commit(): void;
+		/** New in IndexedDB 3.0 - not widely implemented */
+		readonly durability: IDBTransactionDurability;
+    }
+
+	interface IDBCursor {
+		/** New in IndexedDB 3.0 - not widely implemented */
+		readonly request: IDBRequest;
 	}
 }
 
-export type PDBTransactionHandler<T> = (tx: IDBTransaction, helpers: PDBTransactionHelpers) => Promise<T> | T;
 export type PDBUpgradeHandler = (db: IDBDatabase, fromVersion: number, toVersion: number) => void;
 export type PDBMigrationHandler = (db: IDBDatabase) => void;
 
-export type PDBTransactionMode = "readonly" | "readwrite";
+export type PDBTransactionHandler<T> = (tx: IDBTransaction, helpers: PDBTransactionHelpers) => Promise<T> | T;
 
 export interface PDBCursorOptions {
 	range?: IDBKeyRange | IDBValidKey;
-	direction?: PDBCursorDirection;
+	direction?: IDBCursorDirection;
 };
 
 export interface PDBTransactionHelpers {
@@ -34,8 +56,6 @@ export interface PDBTransactionHelpers {
 	/** Configure a timeout for this transaction. If the transaction does not complete within the specified time it will reject with a TimeoutError */
 	timeout: (ms: number) => void;
 }
-
-export type PDBCursorDirection = "next" | "prev" | "nextunique" | "prevunique";
 
 export interface PDBCursor<C extends IDBCursor> {
 	/**